refactor(dbless Part D): ranking パッケージの D1 query コードを全削除
R2 リーダ (read-ranking-items-snapshot 等) が runtime/生成器を賄うため、D1 metrics/sources を
引く旧コードを削除:
- ranking-item D1 repo 15 (find/list/upsert/update/delete 系)
- ranking-tag D1 (get-tags-for-item / sync-ranking-tags)、survey D1 (find-survey-by-id / list-surveys)
- services/sync-ranking-export (D1 updateRankingItem 書込)、scripts/auto-attach-normalization
- 各 index.ts を R2 リーダのみに整理、server.ts から D1 export 除去
- 旧 D1 mock の obsolete テスト 8件削除
- CategoryRankingItem 型を types/ranking-item へ relocate (per-url / 基盤1 が staying 使用)
- calculate-ranking-values / compute-normalization の既存テストは mock を
  readRankingItemByKeyFromR2 に rename して維持 (15 tests green)

検証: ranking / apps/web tsc 0 error。vitest 2ファイル 15 tests 緑。runtime は R2 item.json のみ。

対応: docs/.../quiet-forging-flurry.md Part D (ranking)
