--- packages/ranking/src/repositories/ranking-item/find-ranking-item-by-key.ts
@@ -1,27 +1,0 @@
-import "server-only";
-
-import { getDrizzle, metrics } from "@stats47/database/server";
-import { err, ok, type Result } from "@stats47/types";
-import { eq } from "drizzle-orm";
-import type { RankingItem } from "../../types";
-import { metricAsRankingItemSelection } from "../shared/metric-as-ranking-item-selection";
-import { parseMetricAsRankingItem } from "../shared/parse-metric-as-ranking-item";
-
-export async function findRankingItemByKey(
-  rankingKey: string,
-  db?: ReturnType<typeof getDrizzle>
-): Promise<Result<RankingItem | null, Error>> {
-  try {
-    const drizzleDb = db ?? getDrizzle();
-    const result = await drizzleDb
-      .select(metricAsRankingItemSelection)
-      .from(metrics)
-      .where(eq(metrics.key, rankingKey))
-      .limit(1);
-
-    if (result.length === 0) return ok(null);
-    return ok(parseMetricAsRankingItem(result[0]));
-  } catch (error) {
-    return err(error instanceof Error ? error : new Error(String(error)));
-  }
-}
