--- packages/ranking/src/repositories/ranking-item/list-ranking-items-by-area-type.ts
@@ -1,49 +1,0 @@
-import "server-only";
-
-import { listMetricKeysByEntity } from "@stats47/data-configs";
-import { getDrizzle, metrics } from "@stats47/database/server";
-import { err, ok, type Result } from "@stats47/types";
-import type { AreaType } from "@stats47/types";
-import { and, eq, inArray } from "drizzle-orm";
-import type { RankingItem } from "../../types";
-import { metricAsRankingItemSelection } from "../shared/metric-as-ranking-item-selection";
-import { parseMetricAsRankingItem } from "../shared/parse-metric-as-ranking-item";
-
-export async function listRankingItemsByAreaType(
-  areaType: AreaType,
-  options?: { dataSourceId?: string; categoryKey?: string },
-  db?: ReturnType<typeof getDrizzle>
-): Promise<Result<RankingItem[], Error>> {
-  try {
-    const drizzleDb = db ?? getDrizzle();
-
-    // Phase 7: D1 stats_prefecture EXISTS フィルタを TS-config (`entities`) 由来に置換。
-    if (areaType === "national") {
-      return ok([]);
-    }
-    const allowedKeys = listMetricKeysByEntity(areaType);
-    if (allowedKeys.length === 0) {
-      return ok([]);
-    }
-
-    const conditions = [
-      inArray(metrics.key, allowedKeys),
-      eq(metrics.isActive, true),
-    ];
-    if (options?.categoryKey) conditions.push(eq(metrics.categoryKey, options.categoryKey));
-
-    const result = await drizzleDb
-      .select(metricAsRankingItemSelection)
-      .from(metrics)
-      .where(and(...conditions))
-      .orderBy(metrics.title);
-
-    const items = result
-      .map((row) => { try { return parseMetricAsRankingItem(row); } catch { return null; } })
-      .filter((item): item is RankingItem => item !== null);
-
-    return ok(items);
-  } catch (error) {
-    return err(error instanceof Error ? error : new Error(String(error)));
-  }
-}
